+++ app/Policies/DevicePolicy.php
@@ -32,7 +32,7 @@ class DevicePolicy
      */
     public function view(User $user, Device $device)
     {
-        return $user->is_admin || $device->is_public || groups()->userHasModel($user, $device);
+        return $user->is_admin || $device->is_public || users()->hasIndirectRelationship($user, $device);
     }
 
     /**
@@ -58,9 +58,9 @@ class DevicePolicy
     {
         return $user->is_admin ||
             (
-                tests()->isProperty($device, $user) &&
-                groups()->userIsGroupAdmin($user, $group) &&
-                !groups()->modelIsGroupModel($device, $group)
+                users()->isProperty($device, $user) &&
+                groups()->hasAdministrator($group, $user) &&
+                !resources()->hasGroup($device, $group)
             );
     }
 
@@ -76,9 +76,9 @@ class DevicePolicy
     {
         return $user->is_admin ||
             (
-                tests()->isProperty($device, $user) &&
-                groups()->userIsGroupAdmin($user, $group) &&
-                groups()->modelIsGroupModel($device, $group)
+                users()->isProperty($device, $user) &&
+                groups()->hasAdministrator($group, $user) &&
+                resources()->hasGroup($device, $group)
             );
     }
 
@@ -94,7 +94,7 @@ class DevicePolicy
     {
         return $user->is_admin ||
             (
-                tests()->isProperty($device, $user) &&
+                users()->isProperty($device, $user) &&
                 !in_array($device->uuid, $geofence->devices->pluck('uuid')->toArray())
             );
     }
@@ -111,7 +111,7 @@ class DevicePolicy
     {
         return $user->is_admin ||
             (
-                tests()->isProperty($device, $user) &&
+                users()->isProperty($device, $user) &&
                 in_array($device->uuid, $geofence->devices->pluck('uuid')->toArray())
             );
     }
@@ -125,7 +125,7 @@ class DevicePolicy
      */
     public function update(User $user, Device $device)
     {
-        return $user->is_admin || tests()->isProperty($device, $user);
+        return $user->is_admin || users()->isProperty($device, $user);
     }
 
     /**
@@ -137,6 +137,6 @@ class DevicePolicy
      */
     public function delete(User $user, Device $device)
     {
-        return $user->is_admin || tests()->isProperty($device, $user);
+        return $user->is_admin || users()->isProperty($device, $user);
     }
 }
