+++ app/Policies/PetPolicy.php
@@ -96,9 +96,9 @@ class PetPolicy
     {
         return $user->is_admin ||
             (
-                tests()->isProperty($pet, $user) &&
-                groups()->userIsGroupAdmin($user, $group) &&
-                !groups()->modelIsGroupModel($pet, $group)
+                users()->isProperty($pet, $user) &&
+                groups()->hasAdministrator($group, $user) &&
+                !resources()->hasGroup($pet, $group)
             );
     }
 
@@ -114,9 +114,9 @@ class PetPolicy
     {
         return $user->is_admin ||
             (
-                tests()->isProperty($pet, $user) &&
-                groups()->userIsGroupAdmin($user, $group) &&
-                groups()->modelIsGroupModel($pet, $group)
+                users()->isProperty($pet, $user) &&
+                groups()->hasAdministrator($group, $user) &&
+                resources()->hasGroup($pet, $group)
             );
     }
 
@@ -132,7 +132,7 @@ class PetPolicy
     {
         return $user->is_admin ||
             (
-                tests()->isProperty($pet, $user) &&
+                users()->isProperty($pet, $user) &&
                 !in_array($pet->uuid, $geofence->pets->pluck('uuid')->toArray())
             );
     }
@@ -149,7 +149,7 @@ class PetPolicy
     {
         return $user->is_admin ||
             (
-                tests()->isProperty($pet, $user) &&
+                users()->isProperty($pet, $user) &&
                 in_array($pet->uuid, $geofence->pets->pluck('uuid')->toArray())
             );
     }
