+++ app/Repositories/GroupsRepository.php
@@ -2,13 +2,14 @@
 
 namespace App\Repositories;
 
+use App\Asserts\GroupAssert;
 use App\Group;
-use App\User;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
-use Illuminate\Database\Eloquent\Model;
 
 class GroupsRepository
 {
+    use GroupAssert;
+
     public function getUserGroups()
     {
         return Group::with('user', 'photo')
@@ -59,8 +60,7 @@ class GroupsRepository
             'photo_uuid' => $photo->uuid,
             'photo_url' => config('storage.path') . $photo->photo_url,
         ]);
-        authenticated()
-            ->groups()
+        authenticated()->groups()
             ->attach(
                 $group->uuid,
                 [
@@ -72,77 +72,5 @@ class GroupsRepository
         $group->photos()->attach($photo->uuid);
         $group->load('photo', 'user');
         return $group;
-    }
-
-    /**
-     * Check if user has model.
-     *
-     * @param User $user
-     * @param object|Model $model
-     * @return boolean
-     */
-    public static function userHasModel(User $user, Model $model)
-    {
-        return $model->user_uuid === $user->uuid || $user->groups()
-                ->whereIn(
-                    'uuid',
-                    $model->groups->pluck('uuid')->toArray()
-                )->exists();
-    }
-
-    /**
-     * Check if user is group admin.
-     *
-     * @param object|User $user
-     * @param object|Group $group
-     * @return bool
-     */
-    public static function userIsGroupAdmin(User $user, Group $group)
-    {
-        return $user->uuid === $group->user_uuid || $user->groups()->whereIn('status', [
-            Group::ACCEPTED_STATUS, Group::ATTACHED_STATUS
-        ])->where('is_admin', true)->exists();
-    }
-
-    /**
-     * Check if model is group model.
-     *
-     * @param object|Model $model
-     * @param object|Group $group
-     * @return bool
-     */
-    public static function modelIsGroupModel(Model $model, Group $group)
-    {
-        return $model->groups()->where('group_uuid', $group->uuid)
-            ->whereIn('status', [
-                Group::ACCEPTED_STATUS, Group::ATTACHED_STATUS
-            ])->exists();
-    }
-
-    /**
-     * Check if model is blocked.
-     *
-     * @param object|Model $model
-     * @param object|Group $group
-     * @return bool
-     */
-    public static function modelIsBlocked(Model $model, Group $group)
-    {
-        return $model->groups()->where('group_uuid', $group->uuid)
-            ->whereIn('status', [
-                Group::BLOCKED_STATUS
-            ])->exists();
-    }
-
-    /**
-     * User has a group id.
-     *
-     * @param User $user
-     * @param $groupId
-     * @return bool
-     */
-    public static function userHasGroup(User $user, $groupId)
-    {
-        return in_array($groupId, $user->groups->pluck('uuid')->toArray());
     }
 }
