+++ app/Repositories/ModelLocationsRepository.php
@@ -7,6 +7,7 @@ namespace App\Repositories;
 use App\Contracts\Resource;
 use App\Device;
 use App\Pet;
+use App\Procedures\ResourceLocationProcedure;
 use App\Queries\LocationQuery;
 use App\User;
 use Grimzy\LaravelMysqlSpatial\Eloquent\Builder;
@@ -16,6 +17,7 @@ use Illuminate\Support\Str;
 class ModelLocationsRepository
 {
     use LocationQuery;
+    use ResourceLocationProcedure;
 
     /**
      * Update current model location.
@@ -86,56 +88,5 @@ class ModelLocationsRepository
                 "accuracy" => $data["location"]["coords"]["accuracy"],
             ]);
         }
-    }
-
-    /**
-     * Make filterable locations query.
-     *
-     * @param $models
-     * @param $parameters
-     * @return Builder
-     */
-    public function filterLocations($models, $parameters)
-    {
-        $class = finder()->findClass($parameters['type']);
-        return $this->resolveLocations(
-            $models,
-            $parameters,
-            get_class($class),
-            $class::DEFAULT_LOCATION_DATE_COLUMN,
-            $class::DEFAULT_LOCATION_GROUP_BY_COLUMN
-        );
-    }
-
-    /**
-     * Resolve model locations.
-     *
-     * @param $models
-     * @param $parameters
-     * @param $type
-     * @param $dateColumn
-     * @param $groupedBy
-     * @return Builder
-     */
-    public function resolveLocations($models, $parameters, $type, $dateColumn, $groupedBy)
-    {
-        return $this->groupByColumn(
-            $this->orderByColumn(
-                $this->queryRangeOfDates(
-                    $this->maxAccuracy(
-                        $this->trackableQuery(
-                            $models,
-                            $type
-                        ),
-                        $parameters['accuracy']
-                    ),
-                    $parameters['start_date'],
-                    $parameters['end_date'],
-                    $dateColumn
-                ),
-                $dateColumn
-            ),
-            $groupedBy
-        );
     }
 }
